fixed up handshake, and username, msg sending and recv

+++ client.h
@@ -29,8 +29,8 @@ char *inputMessage(FILE* fp, size_t size);
 
 bool receivedHandshake(int s);
 
-void getStringFromRecv(int s, char * str, int len);
-void receiveMessage(int s, int flag);
+void getStringFromRecv(int s, char * str, uint8_t len);
+void receiveMessage(int s, uint8_t flag);
 
 // void getCurrentUserList(int s, struct username * users, int numberOfUsers);
